Mostly rename tests to have them discovered by pytest
